requirements.txt
python-telegram-bot>=20.0
requests>=2.25.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0